ref(api): replace api service to domain gateways;
